requirements.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==4.2
django-cuser==2017.3.16
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-modeltranslation==0.18.11
django-parler==2.3
django-phone-field==1.8.1
django-tinymce==3.7.1
exceptiongroup==1.2.0
executing==2.0.1
GDAL==3.8.3
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
